requirements.txt
fastapi
uvicorn
python-socketio
httpx
jinja2
